Move React state setter usages from api submission functions back into React components.
